Clean up lsp code a bit
