+++ mypy.ini
@@ -55,6 +55,9 @@ ignore_missing_imports = True
 [mypy-peewee]
 ignore_missing_imports = True
 
+[mypy-pylsp_jsonrpc.*]
+ignore_missing_imports = True
+
 [mypy-semgrep.rule_match]
 disallow_any_decorated = False
 warn_return_any = False
